rewrote validator for backend
